QuizDown Episode 5
Creating an AnswerService to be used by the QuizController to track
the answers from the quiz taker. Uses the (to be written) QuestionStore
which will be responsible for assigning the question number so it can
be looked up later by the AnswerService to associate the response to the
correct question.