QuizDown Episode 6
Next: Completing changes to QuizParser to return a populated
QuestionStore. In the future, we may no longer need an iterator
around the list of questions, we can just use the question number
to look it up in the QuestionStore, and have some way to ensure
that all the questions have been asked.