QuizDown Episode 11
Fixed bug where we dropped the @RequestParam annotation for the form
answer map, which caused us to get the Model's map instead (not what
we wanted).

Fixed bug for fill-in-the-blank answers were not converted to
lower case, and therefore not matching the list of correct choices.

Added a redirect in between the quiz being done and seeing the results
so we can just see results for a specific user.